+++ noodles-bam/src/writer/sam_record.rs
@@ -71,8 +71,7 @@ where
     write_mapping_quality(writer, record.mapping_quality())?;
 
     // bin
-    let alignment_end = record.alignment_end().transpose()?;
-    write_bin(writer, record.position(), alignment_end)?;
+    write_bin(writer, record.position(), record.alignment_end())?;
 
     writer.write_u16::<LittleEndian>(n_cigar_op)?;