+++ noodles-cram/src/data_container/slice.rs
@@ -277,7 +277,7 @@ fn resolve_mates(records: &mut [Record]) -> io::Result<()> {
         let mate = &mut left[i];
         set_mate(record, mate);
 
-        let template_size = calculate_template_size(record, mate)?;
+        let template_size = calculate_template_size(record, mate);
         record.template_size = template_size;
         mate.template_size = -template_size;
 
@@ -307,22 +307,15 @@ fn set_mate(mut record: &mut Record, mate: &mut Record) {
 }
 
 // _Sequence Alignment/Map Format Specification_ (2021-06-03) § 1.4.9 "TLEN"
-fn calculate_template_size(record: &Record, mate: &Record) -> io::Result<i32> {
+fn calculate_template_size(record: &Record, mate: &Record) -> i32 {
     let start = if record.bam_flags().is_reverse_complemented() {
-        record
-            .alignment_end()
-            .transpose()?
-            .map(i32::from)
-            .unwrap_or_default()
+        record.alignment_end().map(i32::from).unwrap_or_default()
     } else {
         record.alignment_start().map(i32::from).unwrap_or_default()
     };
 
     let end = if mate.bam_flags().is_reverse_complemented() {
-        mate.alignment_end()
-            .transpose()?
-            .map(i32::from)
-            .unwrap_or_default()
+        mate.alignment_end().map(i32::from).unwrap_or_default()
     } else {
         mate.alignment_start().map(i32::from).unwrap_or_default()
     };
@@ -335,9 +328,9 @@ fn calculate_template_size(record: &Record, mate: &Record) -> io::Result<i32> {
     // rightmost, and the sign for any middle segment is undefined. If segments cover the same
     // coordinates then the choice of which is leftmost and rightmost is arbitrary..."
     if start > end {
-        Ok(-len)
+        -len
     } else {
-        Ok(len)
+        len
     }
 }
 
@@ -420,7 +413,7 @@ mod tests {
     }
 
     #[test]
-    fn test_calculate_template_size() -> Result<(), Box<dyn std::error::Error>> {
+    fn test_calculate_template_size() -> Result<(), sam::record::position::TryFromIntError> {
         use sam::record::{Flags, Position};
 
         // --> -->
@@ -434,8 +427,8 @@ mod tests {
             .set_read_length(50)
             .build();
 
-        assert_eq!(calculate_template_size(&record, &mate)?, 101);
-        assert_eq!(calculate_template_size(&mate, &record)?, -101);
+        assert_eq!(calculate_template_size(&record, &mate), 101);
+        assert_eq!(calculate_template_size(&mate, &record), -101);
 
         // --> <--
         // This is the example given in _Sequence Alignment/Map Format Specification_ (2021-06-03)
@@ -451,8 +444,8 @@ mod tests {
             .set_read_length(50)
             .build();
 
-        assert_eq!(calculate_template_size(&record, &mate)?, 150);
-        assert_eq!(calculate_template_size(&mate, &record)?, -150);
+        assert_eq!(calculate_template_size(&record, &mate), 150);
+        assert_eq!(calculate_template_size(&mate, &record), -150);
 
         // <-- -->
         let record = Record::builder()
@@ -466,8 +459,8 @@ mod tests {
             .set_read_length(50)
             .build();
 
-        assert_eq!(calculate_template_size(&record, &mate)?, 52);
-        assert_eq!(calculate_template_size(&mate, &record)?, -52);
+        assert_eq!(calculate_template_size(&record, &mate), 52);
+        assert_eq!(calculate_template_size(&mate, &record), -52);
 
         // <-- <--
         let record = Record::builder()
@@ -482,8 +475,8 @@ mod tests {
             .set_read_length(50)
             .build();
 
-        assert_eq!(calculate_template_size(&record, &mate)?, 101);
-        assert_eq!(calculate_template_size(&mate, &record)?, -101);
+        assert_eq!(calculate_template_size(&record, &mate), 101);
+        assert_eq!(calculate_template_size(&mate, &record), -101);
 
         Ok(())
     }